requirements.txt
fastapi
uvicorn
torch
torchvision
pennylane
ultralytics
pillow
opencv-python
numpy